refactor: Car 클래스 분리 및 자동차 객체 생성
