+++ src/main/java/racingcar/Application.java
@@ -1,13 +1,15 @@
 package racingcar;
 
+import model.Car;
+
 import java.util.List;
-import java.util.Set;
 
 public class Application {
     public static void main(String[] args) {
-        // TODO: 프로그램 구현
-        Race race = new Race();
         SetRace set = new SetRace();
-        System.out.println(race.raceCar(set.createCar(),set.inputMovement()));
+        List<Car> cars = set.createCar();
+        for (Car car : cars) {
+            System.out.println(car.getName());
+        }
     }
 }
